+++ ep/ep_hash.h
@@ -17,13 +17,13 @@
 typedef struct EP_HASH		EP_HASH;
 
 typedef void	(*EP_HASH_FORALL_FUNCP)(
-			int keylen,
+			size_t keylen,
 			const void *key,
 			void *val,
 			va_list av);
 typedef int	(*EP_HASH_HASH_FUNCP)(
 			const EP_HASH *const hash,
-			const int keylen,
+			const size_t keylen,
 			const void *key);
 
 extern EP_HASH	*ep_hash_new(
@@ -34,16 +34,16 @@ extern void	ep_hash_free(
 			EP_HASH *hash);
 extern void	*ep_hash_search(
 			EP_HASH *hash,
-			int keylen,
+			size_t keylen,
 			const void *key);
 extern void	*ep_hash_insert(
 			EP_HASH *hash,
-			int keylen,
+			size_t keylen,
 			const void *key,
 			void *val);
 extern void	*ep_hash_delete(
 			EP_HASH *hash,
-			int keylen,
+			size_t keylen,
 			const void *key);
 extern void	ep_hash_forall(
 			EP_HASH *hash,
